Style ajouté au login & register
