+++ DefenseByNight/App_Start/AutoMapperConfig.cs
@@ -1,10 +1,6 @@
 ﻿using AutoMapper;
 using DAL.Mapping;
 using DefenseByNight.Mapping;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
 
 namespace DefenseByNight.App_Start
 {
