+++ DefenseByNight/Areas/AuthentificationManager/Content/custom.css
@@ -1,4 +1,4 @@
-﻿#loginPageBody {
+﻿body {
     background-image: url("../../../Content/Images/Background-Connexion.jpg");
 }
 
@@ -8,22 +8,8 @@
     text-align: center;
 }
 
-.outer {
-    display: table;
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-}
-
-.middle {
-    display: table-cell;
-    vertical-align: middle;
-}
-
-.inner {
-    margin-left: auto;
-    margin-right: auto;
-    width: 500px;
+.center {
+    text-align: center;
+    margin-left: 20%;
+    margin-right: 20%
 }
