+++ DefenseByNight/Areas/AuthentificationManager/Controllers/AuthController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNet.Identity;
 using System.Threading.Tasks;
 using Microsoft.Owin.Security;
 using DAL.Models.Identity;
+using DefenseByNight.Controllers;
 
 namespace DefenseByNight.Areas.AuthentificationManager.Controllers
 {
