--- DefenseByNight/Areas/AuthentificationManager/Views/Auth/_Connexion.cshtml
@@ -1,35 +1,0 @@
-﻿@model UserConnexionViewModel
-
-<div id="ErrLoginFailed" class="alert alert-danger" role="alert" hidden></div>
-
-@using (Ajax.BeginForm("SignIn", "Login", new { area = "AuthentificationManager" }, new AjaxOptions { HttpMethod = "POST", OnSuccess = "LoginModule.SignIn(data);" }
-        , new { @class = "form-horizontal", id = "connexionForm" }))
-{
-
-    @Html.AntiForgeryToken()
-    HtmlHelper.UnobtrusiveJavaScriptEnabled = false;
-
-    <div class="form-group">
-        @Html.LabelFor(m => m.Email, new { @class = "col-sm-3 control-label", id = "res_email_connexion_lbl" })
-        <div class="col-sm-9">
-            @Html.TextBoxFor(m => m.Email, new { id = "EmailConnexion", @class = "form-control", @type = "email", placeholder = "monEmail@trucmachin.com" })
-            @Html.ValidationMessageFor(m => m.Email, "", new { @class = "text-danger " })
-        </div>
-    </div>
-
-    <div class="form-group">
-        @Html.LabelFor(m => m.Password, new { @class = "col-sm-3 control-label", id = "res_password_connexion_lbl" })
-        <div class="col-sm-9">
-            @Html.TextBoxFor(m => m.Password, new { id = "PasswordConnexion", @class = "form-control", @type = "password", placeholder = "******" })
-            @Html.ValidationMessageFor(m => m.Password, "", new { @class = "text-danger " })
-        </div>
-    </div>
-
-
-
-    <div class="form-group">
-        <div class="text-center">
-            <button id="submitButton" type="submit" class="btn btn-primary"></button>
-        </div>
-    </div>
-}
