--- DefenseByNight/Areas/AuthentificationManager/Views/Auth/_Register2.cshtml
@@ -1,70 +1,0 @@
-﻿@model UserRegistrationViewModel
-
-@using (Ajax.BeginForm("SignUp", "Login", new { area = "AuthentificationManager" }, new AjaxOptions { HttpMethod = "POST" }
-    , new { @class = "form-horizontal", id = "registrationForm" }))
-{
-    {
-        @Html.AntiForgeryToken()
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.FirstName, htmlAttributes: new { id = "res_firstname_registration_lbl", @class = "control-label col-md-3" })
-            <div class="col-md-9">
-                @Html.EditorFor(model => model.FirstName, new { htmlAttributes = new { @class = "form-control", placeholder = "Pierre" } })
-                @Html.ValidationMessageFor(model => model.FirstName, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.LastName, htmlAttributes: new { id = "res_lastname_registration_lbl", @class = "control-label col-md-3" })
-            <div class="col-md-9">
-                @Html.EditorFor(model => model.LastName, new { htmlAttributes = new { @class = "form-control", placeholder = "Paul" } })
-                @Html.ValidationMessageFor(model => model.LastName, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.Alias, htmlAttributes: new { id = "res_alias_registration_lbl", @class = "control-label col-md-3" })
-            <div class="col-md-9">
-                @Html.EditorFor(model => model.Alias, new { htmlAttributes = new { @class = "form-control", placeholder = "Houjack" } })
-                @Html.ValidationMessageFor(model => model.Alias, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.Email, htmlAttributes: new { id = "res_email_registration_lbl", @class = "control-label col-md-3" })
-            <div class="col-md-9">
-                @Html.EditorFor(model => model.Email, new { htmlAttributes = new { @class = "form-control", placeholder = "MonEmail@trucmachin.com" } })
-                @Html.ValidationMessageFor(model => model.Email, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.Password, htmlAttributes: new { id = "res_password_registration_lbl", @class = "control-label col-md-3" })
-            <div class="col-md-9">
-                @Html.EditorFor(model => model.Password, new { htmlAttributes = new { @class = "form-control", placeholder = "******", @type = "password" } })
-                @Html.ValidationMessageFor(model => model.Password, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.Password, htmlAttributes: new { id = "res_password2_registration_lbl", @class = "control-label col-md-3" })
-            <div class="col-md-9">
-                @Html.TextBox("ConfirmPassword", "", new { @class = "form-control", placeholder = "******", @type = "password" })
-                @Html.ValidationMessage("ConfirmPassword", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="form-group">
-            @Html.LabelFor(model => model.BirthDate, htmlAttributes: new { id = "res_birthdate_registration_lbl", @class = "control-label col-md-3" })
-            <div class="col-md-9">
-                @Html.EditorFor(model => model.BirthDate, new { htmlAttributes = new { id = "birthdate_registration", @class = "form-control" } })
-                @Html.ValidationMessageFor(model => model.BirthDate, "", new { @class = "text-danger" })
-            </div>
-        </div>
-
-        <div class="text-center">
-            <input id="createUser" type="submit" class="btn btn-primary" />
-        </div>
-    }
-}
-
