+++ DefenseByNight/packages.config
@@ -1,13 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Antlr" version="3.4.1.9004" targetFramework="net47" />
+  <package id="Antlr" version="3.5.0.2" targetFramework="net47" />
   <package id="Autofac" version="4.9.2" targetFramework="net47" />
   <package id="Autofac.Mvc5" version="4.0.2" targetFramework="net47" />
   <package id="AutoMapper" version="8.0.0" targetFramework="net47" />
   <package id="bootstrap" version="3.4.1" targetFramework="net47" />
   <package id="EntityFramework" version="6.2.0" targetFramework="net47" />
   <package id="FontAwesome" version="4.7.0" targetFramework="net47" />
-  <package id="jQuery" version="3.3.1" targetFramework="net47" />
+  <package id="jQuery" version="3.4.1" targetFramework="net47" />
   <package id="jquery.validate.bootstrap" version="1.1.0" targetFramework="net47" />
   <package id="jQuery.Validation" version="1.17.0" targetFramework="net47" />
   <package id="Microsoft.AspNet.Identity.Core" version="2.2.2" targetFramework="net47" />
@@ -23,13 +23,13 @@
   <package id="Microsoft.Net.Compilers" version="3.0.0" targetFramework="net47" developmentDependency="true" />
   <package id="Microsoft.Owin" version="4.0.1" targetFramework="net47" />
   <package id="Microsoft.Owin.Host.SystemWeb" version="4.0.1" targetFramework="net47" />
-  <package id="Microsoft.Owin.Security" version="3.0.1" targetFramework="net47" />
-  <package id="Microsoft.Owin.Security.Cookies" version="3.0.1" targetFramework="net47" />
-  <package id="Microsoft.Owin.Security.OAuth" version="3.0.1" targetFramework="net47" />
+  <package id="Microsoft.Owin.Security" version="4.0.1" targetFramework="net47" />
+  <package id="Microsoft.Owin.Security.Cookies" version="4.0.1" targetFramework="net47" />
+  <package id="Microsoft.Owin.Security.OAuth" version="4.0.1" targetFramework="net47" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net47" />
-  <package id="Modernizr" version="2.6.2" targetFramework="net47" />
+  <package id="Modernizr" version="2.8.3" targetFramework="net47" />
   <package id="Moment.js" version="2.24.0" targetFramework="net47" />
-  <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net47" />
+  <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net47" />
   <package id="Owin" version="1.0" targetFramework="net47" />
   <package id="System.ValueTuple" version="4.5.0" targetFramework="net47" />
   <package id="WebGrease" version="1.5.2" targetFramework="net47" />
